fix: correct watch clock drift on time sync from phone
Two issues in Bangle::setTime() combined to make the watch's clock
drift from the phone's by up to a minute or more on every sync:
timeOffset was declared float, so timeUnix + timeOffset*3600 got
computed in single-precision float, rounding a ten-digit Unix
timestamp to the nearest multiple of ~128 seconds; on top of that, an
undocumented +20 second constant (added upstream with no explanation)
was tacked onto every sync. Widening timeOffset to double removes the
rounding error, and dropping the unexplained +20 removes the rest.
Verified live: computed sync values now match the phone's raw
timestamp exactly, with no rounding artifacts.

+++ main/src/Notifs/Bangle.h
@@ -55,7 +55,7 @@ private:
 
 	std::unordered_set<uint32_t> missedCalls;
 
-	float timeOffset = 0;
+	double timeOffset = 0;
 	uint64_t timeUnix = 0;
 
 	void setTime();
